requirements.txt
langchain
langchain-anthropic
langchain-community
tavily-python
python-dotenv
